requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
psycopg2-binary==2.9.9
PySocks==1.7.1
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
schedule==1.2.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
spotipy==2.23.0
telebot==0.0.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
wsproto==1.2.0
